backend/requirements.txt
# ScriptToVideo AI Pipeline — Backend Dependencies
# Install: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Task Queue ─────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.0.4
billiard>=4.2.0              # Required for Celery solo pool on Windows

# ── AI Services ────────────────────────────────────────────────────────────────
openai==1.30.0               # GPT-4o scene breakdown + metadata
httpx==0.27.0                # ElevenLabs TTS HTTP calls

# ── TTS Fallback ───────────────────────────────────────────────────────────────
gTTS==2.5.1                  # Free Google TTS fallback

# ── Media Processing ───────────────────────────────────────────────────────────
Pillow>=11.0.0               # Image generation utilities
pydub==0.25.1                # Audio processing utilities
python-docx==1.1.0           # .docx script file support

# ── YouTube Upload ─────────────────────────────────────────────────────────────
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ── Authentication ─────────────────────────────────────────────────────────────
PyJWT>=2.8.0                 # JWT token encode/decode

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.30